refactor bo
